+++ src/lte/model/lte-ue-mac.cc
@@ -841,7 +841,7 @@ LteUeMac::AssignStreams (int64_t stream)
   return 1;
 }
 
-void LteUeMac::SendCognitiveMessage(Ptr<SpectrumSignalParameters> rxParams, uint32_t UnexpectedAccessBitmap, bool falsePositive)
+void LteUeMac::SendCognitiveMessage(Ptr<SpectrumSignalParameters> rxParams, uint64_t UnexpectedAccessBitmap, bool falsePositive)
 {
     //We stopped receiving ctrl messages(not supposed to happen), so estimate frames
     //if (lastFrameNo == m_frameNo && lastSubframeNo == m_subframeNo)
